Apply clippy suggestions for Rust 1.66

+++ src/offset/utc.rs
@@ -68,8 +68,7 @@ impl Utc {
         let now =
             SystemTime::now().duration_since(UNIX_EPOCH).expect("system time before Unix epoch");
         let naive =
-            NaiveDateTime::from_timestamp_opt(now.as_secs() as i64, now.subsec_nanos() as u32)
-                .unwrap();
+            NaiveDateTime::from_timestamp_opt(now.as_secs() as i64, now.subsec_nanos()).unwrap();
         DateTime::from_utc(naive, Utc)
     }
